// frontend/src/app/api/model/switch.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Switch { 
    /**
     * The unique identifier of this switch
     */
    readonly id?: number;
    /**
     * The friendly name of this switch
     */
    description: string;
    /**
     * The community string used for SNMP access to this switch
     * @deprecated
     */
    community?: string | null;
    /**
     * The IPv4 address of this switch
     */
    ip: string;
}